Delete translation if there is more than one translation

+++ app/Http/Controllers/NodesController.php
@@ -363,7 +363,11 @@ class NodesController extends ReactorController {
 
         if ($response = $this->validateNodeIsNotLocked($node)) return $response;
 
-        $source->delete();
+        // Prevent deletion if there is only one translation
+        if (count($node->translations()) > 1)
+        {
+            $source->delete();
+        }
 
         $node->load('translations');
